src/main/java/org/example/authentication/utils/PasswordHasher.java
package org.example.authentication.utils;

import javax.crypto.SecretKeyFactory;
import javax.crypto.spec.PBEKeySpec;
import java.security.NoSuchAlgorithmException;
import java.security.spec.InvalidKeySpecException;
import java.security.spec.KeySpec;
import java.util.Base64;

public class PasswordHasher {
    private final static int ITERATION_COUNT = 65536;
    private final static int KEY_LENGTH = 128;

    public static String hashPassword(String password, String login) {
        KeySpec spec = new PBEKeySpec(password.toCharArray(), login.getBytes(), ITERATION_COUNT, KEY_LENGTH);
        SecretKeyFactory factory;
        try {
            factory = SecretKeyFactory.getInstance("PBKDF2WithHmacSHA1");
            byte[] hash = factory.generateSecret(spec).getEncoded();
            return Base64.getEncoder().encodeToString(hash);
        } catch (NoSuchAlgorithmException | InvalidKeySpecException e) {
            throw new RuntimeException(e);
        }
    }

    public static boolean verifyPassword(String enteredPassword, String hashedPasswordFromDatabase){
        return enteredPassword.equals(hashedPasswordFromDatabase);
    }
}